[xaprepare] Fix JDK 1.8 detection
Attempting to parse the `JAVA_VERSION="1.8.0_202"` entry from the
`release` text file in JDK 1.8 was failing as it didn't account for any
underscores.  After fixing this, the properties containing version info
for the on disk and downloadable versions of JDK 1.8 needed updating.
On disk we expect `1.8.0.202`, while the download URI contains `8u202`.