[ci] Update Visual Studio instance before build/test jobs
We should update the VS instance on our Windows machines regularly.
There is an extra incentive to do so now, now that .NET 5 is stable.
Since we will be installing .NET 5 into a system location via VS, we
will also need to bump the .NET version in our "preview" sandbox to make
sure our sandbox SDK installation is still preferred.